Move bursar integration tests into separate folder (#126)
* Update submodule

* Move no fees fines test

* Remove temporary test code in AccountReader

* resolve issues with original pull request :#122

---------
